refactor: address PR review feedback for state verifier
- Fix DST compliance: replace Utc::now() with fixed base timestamp in
  all tests to ensure deterministic behavior per testing.md rules
- Fix .unwrap() in production code: replace with let-else pattern in
  detect_suspected_partition to satisfy code-style.md guidelines
- Split 2800-line state_verifier.rs into submodules for maintainability:
  types.rs (data types), detectors.rs (algorithms), report.rs (display),
  tests.rs (unit tests), mod.rs (StateVerifier + build logic)
- Add integration tests for SimNetwork.verify_state() and filter methods
  exercised through real simulation event logs
